Added Heavy Ship. Changed Explosions. Added Damage Behaviour of Lasers

+++ Game/Assets/Scripts/Laser.cs
@@ -3,8 +3,8 @@ using System.Collections;
 
 public class Laser : MonoBehaviour {
 
-    public float damage;
-    public float speed;
+    public int damage;
+    public int speed;
 
 	// Use this for initialization
 	void Start () {
